Use our paredit for form selection in eval, etc
